Fix backend user service + remove code duplication
