+++ backend/user-module/controllers/userController.ts
@@ -7,8 +7,6 @@ import {
   deleteUser,
   register,
   login,
-  followCreator,
-  unfollowCreator,
   followUser,
   unfollowUser,
   requestFollow,
@@ -100,31 +98,6 @@ export async function deleteUserController(req: Request, res: Response) {
   } catch (error) {
     console.log(error);
     res.status(500).json({ error: "Failed to delete user" });
-  }
-}
-
-// Content View - Follow Creator
-export async function followCreatorController(req: Request, res: Response) {
-  console.log("Following creator...");
-  const { userId, creatorId } = req.params;
-  try {
-    await followCreator(userId, creatorId);
-    res.status(200).json({ message: "Creator followed successfully" });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ error: "Failed to follow creator" });
-  }
-}
-
-// Content View - Unfollow Creator
-export async function unfollowCreatorController(req: Request, res: Response) {
-  const { userId, creatorId } = req.params;
-  try {
-    await unfollowCreator(userId, creatorId);
-    res.status(200).json({ message: "Creator unfollowed successfully" });
-  } catch (error) {
-    console.log(error);
-    res.status(500).json({ error: "Failed to unfollow creator" });
   }
 }
