+++ backend/user-module/models/userModel.ts
@@ -14,7 +14,7 @@ export interface User {
   content?: string[]; // Optional content
   likedContent?: string[]; // Optional liked content
   bookmarkedContent?: string[]; // Optional bookmarked content
-  followedCreators?: string[]; // Optional followed creators
-  followedBy?: string[]; // Optional followed by users
+  following?: string[]; // Optional followed creators
+  followers?: string[]; // Optional followed by users
   sharedContent?: string[]; // Optional shared content
 }
