+++ frontend/src/app/styles/profile/ProfileManagement.scss
@@ -50,6 +50,12 @@ textarea {
   color: var(--text-color);
 }
 
+textarea {
+  resize: vertical;
+  min-height: 50px;
+  max-height: 200px;
+}
+
 .save-button {
   width: 100%;
